Started work on the base Weapon class
decided against trying to implement interfaces, but weapon class does allow for an override if there is a special weapon. E.G; make a change to how the weapon primary fires other than default, make a new script for it that inherits from weapon, and override it as so. The rest of the class should work fine even after an external override, and requires a little bit of a converter if there is a special weapon class.