All weapon systems finished
Every weapon in the game currently works (With strange bugs, don't worry)

They are sort of hard to use!! because there is no responsivity. BUT! The key elements do work. I will continue work to make sure they are solid.

There is an example item pickup, viewmodel, and script, along with each of their scriptable objects, for every weapon. Test it out, if you dare!